qcow2: avoid nb_sectors int overflow at zero/discard
during discard, signed integer overflow leads to end_offset possibly
be less than offset,
that in turn makes nb_clusters (to discard) much bigger than expected,
leading to a possible data loss, operation taking very long and
image growing large as discard will allocate L2 tables for all that
huge amount of clusters

One possibility where nb_sectors is large enough (>=2^21) is vm_state
discard at snapshot creation.

Fixed on master in 2.10 with:

    d2cb36af2b0040d421b347e6e4e803e07220f78d
    Author: Eric Blake <eblake@redhat.com>
    Date:   Sat May 6 19:05:52 2017 -0500

    qcow2: Discard/zero clusters by byte count

+++ block/qcow2-cluster.c
@@ -1510,7 +1510,7 @@ static int discard_single_l2(BlockDriverState *bs, uint64_t offset,
 }
 
 int qcow2_discard_clusters(BlockDriverState *bs, uint64_t offset,
-    int nb_sectors, enum qcow2_discard_type type, bool full_discard)
+    uint64_t nb_sectors, enum qcow2_discard_type type, bool full_discard)
 {
     BDRVQcow2State *s = bs->opaque;
     uint64_t end_offset;
@@ -1591,8 +1591,8 @@ static int zero_single_l2(BlockDriverState *bs, uint64_t offset,
     return nb_clusters;
 }
 
-int qcow2_zero_clusters(BlockDriverState *bs, uint64_t offset, int nb_sectors,
-                        int flags)
+int qcow2_zero_clusters(BlockDriverState *bs, uint64_t offset,
+                        uint64_t nb_sectors, int flags)
 {
     BDRVQcow2State *s = bs->opaque;
     uint64_t nb_clusters;
